# contracts/api-contracts/src/main/resources/openapi/signalharvester-v1.yaml
openapi: 3.1.0
info:
  title: SignalHarvester Backend API
  version: 0.1.0
  description: Public HTTP contract for SignalHarvester configuration, results, event observation, processing-flow, and operational
    APIs.
...
security:
- cookieAuth: []
- bearerAuth: []
paths:
...
  /api/v1/admin/operations/changes:
    get:
      tags: [Operational Intelligence]
      operationId: listOperationalChanges
      summary: List recent sanitized operational changes
      parameters:
      - name: limit
        in: query
        required: false
        schema: {type: integer, minimum: 1, maximum: 200, default: 50}
      responses:
        '200':
          description: Recent operational changes
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/OperationalChange'}
        '401': {description: Authentication required}
        '403': {description: ADMIN role required}
  /api/v1/admin/operations/changes/markers:
    post:
      tags: [Operational Intelligence]
      operationId: createOperationalMarker
      summary: Record a bounded operational tooling or scenario marker
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OperationalMarkerRequest'}
      responses:
        '201':
          description: Operational marker recorded
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OperationalChange'}
        '400': {description: Invalid or unsafe operational marker}
        '401': {description: Authentication required}
        '403': {description: ADMIN role required or CSRF validation failed}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: SIGNALHARVESTER_AUTH
      description: HttpOnly JWT cookie used by browser REST and native EventSource requests.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer transport for approved non-browser clients.
...
  schemas:
...
    OperationalChange:
      type: object
      required: [id, changedAt, category, targetType, targetId, beforeState, afterState, outcome, source, actorId, correlationId, traceId, applicationVersion]
      properties:
        id: {type: string, format: uuid}
        changedAt: {type: string, format: date-time}
        category:
          type: string
          enum: [SOURCE_CONFIGURATION, MONITORING_PROFILE_CONFIGURATION, SECURITY_ADMINISTRATION, DEPLOYMENT_TUNING, DEAD_LETTER_RECOVERY, TEST_SCENARIO]
        targetType:
          type: string
          enum: [SOURCE, MONITORING_PROFILE, USER, DEPLOYMENT, DEAD_LETTER_RECORD, SCENARIO]
        targetId: {type: string}
        beforeState:
          type: object
          additionalProperties: {type: string}
        afterState:
          type: object
          additionalProperties: {type: string}
        outcome: {type: string, enum: [APPLIED, REJECTED]}
        source: {type: string, enum: [REST_UI, SYSTEM, TOOLING, TEST_SCENARIO, RECOVERY]}
        actorId: {type: string}
        correlationId: {type: string}
        traceId: {type: string}
        applicationVersion: {type: string}
    OperationalMarkerRequest:
      type: object
      required: [category, targetType, targetId]
      properties:
        category:
          type: string
          enum: [DEPLOYMENT_TUNING, TEST_SCENARIO]
        targetType:
          type: string
          enum: [DEPLOYMENT, SCENARIO]
        targetId: {type: string, minLength: 1, maxLength: 256}
        details:
          type: object
          maxProperties: 32
          additionalProperties: {type: string, maxLength: 512}
...